v2.0.7
Added reference to the assembly reflection that makes it easier for me to change version number strings in a couple of places in the application (i.e., on the splash and about pages).

Fixed a bug where the "load conversions" button was running a switch on the Default Stop List selection rather than the conversions selection, causing a language mismatch when trying to load up conversions.