v2.0.9
Removed logging during the reportprogress method. On systems where there are a lot of threads available, something is going on with how the parallel.foreach block is interacting with the Richtextbox. It was probably foolish to use it this way in the first place, but so it goes. It was throwing a DisconnectedContext error that somehow wasn't being handled by any of the try-catch blocks. I've tried a few things to fix with no success, so I figured that it's better to just let it go for now.