v2.0.10
Just started having issues where the background worker wouldn't move past the last DTM log writing call. The parallel.foreach loop completed without any issues, but something about that log call was breaking. processor would just idle, garbage collection would stop, etc. etc. Very strange.

Tried a few things to fix it, but no success. Ultimately, the problem seems to be better when taking a bit of strain off of the reportprogress method. Specifically, I've made it such that the actual progressbar value isn't updated more than ~1000 times in total. I've also got one last final call to the reportprogress method (seemed to help sometimes).

So, ultimately, I'm not sure what's going on here. I didn't encounter this problem until today (and on a dataset that I've already analyzed twice before), so this is kind of out of left field. The somewhat random occurrence of this issue is also baffling. I'm not confident that the issue is resolved (the only way that it seemed to perfectly resolve was by removing the last log writing call altogether -- weird). But, it seems to be worked-around for time being.