v2.0.12
Test build. Added a new option that uses dictionaries (rather than regular expressions) for the "Conversions" process. For extremely large conversion lists (such as when users are performing lemmatization via conversions rather than using a lemmatizer), this option would be preferable. The use of the regex method for conversions basically caused the program to crash on large conversion lists. Testing with an italian conversion list of about 350,000 items was causing problems, so this method might help to ameliorate the issue. Currently only a test build, so there's no actual build up on the website at the moment.