Restructuring
Restructured the code so that the lemmagen libraries are now part of the project. This also includes their distribution of the lzma library, which previously built as LZMA#.dll. All of the libraries are now build from scratch each time and are part of the overall project, which helps to ensure that they'll all operating under the same version of .NET, among other standardization issues.

Also went in and separated out the TwitterAwareTokenizer so that it is now its own DLL rather than just a class in the project. Never know when you'll want to reuse it for another project.