v2.1.03
Added a validation / sanity check on output options. Previously, it was possible to choose to output one of the document-term matrices without having a frequency list generated as well. Now, it makes sure that the user has the Frequency List type of output selected prior to attempting to extract the DTM output types.